move ExtStateHandler to workflow folder
