+++ roles/lib/files/FWO.Report/ReportGenerator.cs
@@ -6,9 +6,9 @@ using FWO.Data;
 using FWO.Data.Modelling;
 using FWO.Data.Report;
 using FWO.Logging;
-using FWO.Services;
 using FWO.Services.Modelling;
 using FWO.Services.RuleTreeBuilder;
+using FWO.Services.Workflow;
 
 namespace FWO.Report
 {
