+++ roles/middleware/files/FWO.Middleware.Server/Jobs/VarianceAnalysisJob.cs
@@ -8,6 +8,7 @@ using FWO.Logging;
 using FWO.Report;
 using FWO.Services;
 using FWO.Services.Modelling;
+using FWO.Services.Workflow;
 using Quartz;
 
 namespace FWO.Middleware.Server.Jobs
