+++ roles/tests-unit/files/FWO.Test/ModellingVarianceAnalysisTest.cs
@@ -6,6 +6,7 @@ using FWO.Data.Report;
 using FWO.Data.Workflow;
 using FWO.Services;
 using FWO.Services.Modelling;
+using FWO.Services.Workflow;
 using FWO.Basics;
 
 namespace FWO.Test
