+++ roles/ui/files/FWO.UI/Pages/NetworkModelling/NetworkModelling.razor
@@ -2,6 +2,7 @@
 @using FWO.ExternalSystems
 @using FWO.Data
 @using FWO.Ui.Display
+@using FWO.Services.Workflow
 @using RestSharp
 @using System.Net
 @using System.Linq
