+++ roles/ui/files/FWO.UI/Pages/NetworkModelling/RequestFwChangePopup.razor
@@ -1,5 +1,6 @@
 @using FWO.Data.Middleware
 @using FWO.Data.Workflow
+@using FWO.Services.Workflow
 @using FWO.Middleware.Client
 @using FWO.Report
 @using FWO.ExternalSystems.Tufin.SecureChange
