+++ roles/ui/files/FWO.UI/Pages/NetworkModelling/RequestRecertPopup.razor
@@ -6,6 +6,7 @@
 @using FWO.Recert
 @using FWO.Report
 @using FWO.Services.RuleTreeBuilder
+@using FWO.Services.Workflow
 @using FWO.Ui.Pages.Reporting
 @using Newtonsoft.Json
 @using RestSharp
